+++ src/components/SearchField/SearchField.classNames.ts
@@ -6,7 +6,7 @@ import { SearchFieldProps } from './SearchField';
 
 export const getClassNames = (props: SearchFieldProps) => {
   const palette = getTheme().palette as PaletteProps;
-  const { border, searchFieldSize } = props;
+  const { border, searchFieldSize, onSearchIcon } = props;
   const largeSize = searchFieldSize === 'large';
   const standardSize = searchFieldSize === 'standard';
 
@@ -19,18 +19,24 @@ export const getClassNames = (props: SearchFieldProps) => {
       listStyleType: 'none !important',
       padding: '0px',
       marginTop: '0px',
-      marginBottom: '11px'
+      marginBottom: '11px',
+      border: `1px solid ${palette.skeColor.black}`
     },
     searchListDropdown: {
-      width: '100% - 2px',
-      top: '-13px',
-      border: `1px solid ${palette.skeColor.black}`,
+      top: '-1px',
       borderTop: 0,
+      position: 'relative',
       selectors: {
         ul: {
-          margin: 0
+          margin: 0,
+          position: 'absolute',
+          left: 0,
+          top: '100%',
+          width: '99.5%',
+          zIndex: 10
         },
         li: {
+          background: palette.skeColor.white,
           cursor: 'pointer',
           color: palette.skeColor.blackAlt,
           padding: '3px 9px 3px 9px',
@@ -103,7 +109,13 @@ export const getClassNames = (props: SearchFieldProps) => {
           boxSizing: 'border-box',
           fontSize: largeSize ? FontSizes.xLarge : FontSizes.medium,
           opacity: 1,
-          outline: 'none'
+          outline: 'none',
+          cursor: onSearchIcon ? 'pointer' : 'default',
+          selectors: {
+            ':hover': {
+              background: onSearchIcon && palette.skeColor.lightBlue
+            }
+          }
         },
         '& .ms-SearchBox-clearButton': {
           display: 'none'
@@ -115,6 +127,9 @@ export const getClassNames = (props: SearchFieldProps) => {
           WebkitAppearance: 'textfield'
         },
         'input[type=search]::-webkit-search-cancel-button': {
+          WebkitAppearance: 'none'
+        },
+        'input[type="search"]::-webkit-search-decoration': {
           WebkitAppearance: 'none'
         },
         'input[type=search]::-ms-clear': {