+++ src/components/SearchField/SearchField.tsx
@@ -46,6 +46,10 @@ export interface SearchFieldProps extends ISearchBoxProps {
   keyboardShortcut?: boolean;
   /** Hvilke taster som fungerer for snarvei */
   searchShortcutKeys?: 'string';
+  /** Gjør søkeikonet klikkbart, trenger samme kode som onSearch */
+  onSearchIcon?: (newValue: any) => void;
+  /** Legg til egen mouseover tittel på søkeikonet */
+  searchIconTitle?: 'string';
 }
 
 const searchInList = (options: Array<IDropdownOption>, filterText: string) => {
@@ -88,6 +92,8 @@ const SearchField: React.FC<SearchFieldProps> = props => {
     limit,
     keyboardShortcut = false,
     searchShortcutKeys = 'ctrl+f',
+    onSearchIcon,
+    searchIconTitle = 'Søk',
     ...rest
   } = props;
   const _searchBoxElement = React.createRef<HTMLDivElement>();
@@ -263,6 +269,9 @@ const SearchField: React.FC<SearchFieldProps> = props => {
             onKeyDown={ev => handleOnKeyDown(ev)}
             value={value}
             componentRef={_componentRef}
+            iconProps={{
+              onClick: ev => (onSearchIcon ? onSearchIcon(ev) : null)
+            }}
           />
           {dropdownVisible && renderSuggestions(searchResultList)}
         </div>
@@ -272,6 +281,10 @@ const SearchField: React.FC<SearchFieldProps> = props => {
           {...props}
           className={classnames(styles.main, className)}
           componentRef={_componentRef}
+          iconProps={{
+            onClick: ev => (onSearchIcon ? onSearchIcon(ev) : null),
+            title: onSearchIcon ? searchIconTitle : ''
+          }}
         />
       )}
     </div>