+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### **v.3.7.1 - 03.11.2020 **
+
+- Table: Added compact table
+- DatePicker: Fixes a bug where icons where not displayed correctly
+- SearchField: ability to trigger search by clicking icon, drop down menu is now overlayed.
+- Pagination: className is optional as it should be
+- TopStripe: Accessibility and styling fixes.
+- Sketch file updated to v1.5.
+- Updated documentation
+
 ### **v.3.7.0 - 29.09.2020 **
 
 - SearchField: added ability to activate with keyboard shortcut